requirements.txt
django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
polyline==2.0.0
gunicorn==20.1.0